requirements.txt
torch 
transformers 
pillow 
uvicorn 
python-multipart
streamlit
huggingface_hub
accelerate
dotenv